pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pymeshzork"
version = "0.1.0"
description = "Python Zork with GUI Map Editor and Meshtastic Multiplayer"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "haxorthematrix" }
]
keywords = ["zork", "adventure", "game", "meshtastic", "multiplayer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Games/Entertainment :: Role-Playing",
]

dependencies = [
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
gui = [
    "PyQt6>=6.5.0",
]
mesh = [
    "paho-mqtt>=2.0.0",
    "meshtastic>=2.0.0",
]

[project.scripts]
zork = "pymeshzork.cli:main"
zork-editor = "pymeshzork.editor.main:main"

[project.urls]
Homepage = "https://github.com/haxorthematrix/pymeshzork"
Repository = "https://github.com/haxorthematrix/pymeshzork"

[tool.hatch.build.targets.wheel]
packages = ["pymeshzork"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
